refactor module_tree out of being monolithic
