+++ crates/ingest/syn_parser/src/resolve/tests.rs
@@ -1,0 +1,8 @@
+use super::*;
+
+use crate::utils::test_setup::run_phases_and_collect;
+
+#[test]
+fn test_name() {
+    todo!();
+}
