+++ crates/ingest/syn_parser/tests/uuid_phase3_resolution/mod.rs
@@ -1,8 +1,14 @@
-#![cfg(not(feature = "type_bearing_ids"))]
-pub mod canon_resolver;
-pub mod edge_cases;
-pub mod exports; // Add the new module for export tests
-pub mod mod_tree;
 pub mod mod_tree_construction;
+
+#[cfg(not(feature = "type_bearing_ids"))]
+pub mod canon_resolver;
+#[cfg(not(feature = "type_bearing_ids"))]
+pub mod edge_cases;
+#[cfg(not(feature = "type_bearing_ids"))]
+pub mod exports; // Add the new module for export tests
+#[cfg(not(feature = "type_bearing_ids"))]
+pub mod mod_tree;
+#[cfg(not(feature = "type_bearing_ids"))]
 pub mod path_attribute;
+#[cfg(not(feature = "type_bearing_ids"))]
 pub mod shortest_path;
