+++ crates/ingest/syn_parser/tests/uuid_phase3_resolution/mod_tree_construction.rs
@@ -1,3 +1,5 @@
+#![cfg(test)]
+#![allow(unused_imports)]
 //! Tests focusing on the construction and initial state of the ModuleTree,
 //! before path resolution logic is implemented.
 //!
@@ -18,37 +20,25 @@
 use std::collections::HashSet;
 use std::path::Path;
 
+use colored::Colorize as _;
+use itertools::Itertools;
+use log::debug;
+use syn_parser::error::SynParserError;
 use syn_parser::parser::graph::GraphAccess as _;
-use syn_parser::parser::nodes::{GraphId, ModuleNodeId};
-use syn_parser::parser::relations::{Relation, RelationKind};
+use syn_parser::parser::nodes::{ModuleNode, ModuleNodeId};
+use syn_parser::parser::relations::SyntacticRelation;
+use syn_parser::parser::ParsedCodeGraph;
+use syn_parser::resolve::module_tree::ModuleTree;
+use syn_parser::utils::{LogStyle, LOG_TARGET_MOD_TREE_BUILD};
 
 use crate::common::build_tree_for_tests;
-// Removed unused imports for helpers moved to CodeGraph
 
-/// **Covers:** Basic sanity check. Ensures that the number of `ModuleNode`s stored
-/// in the `ModuleTree`'s internal map (`tree.modules()`) matches the total number
-/// of `ModuleNode`s present in the input `CodeGraph`. This confirms that
-/// `ModuleTree::add_module` was called for every module without losing any.
-#[test]
-#[cfg(not(feature = "type_bearing_ids"))]
-fn test_module_tree_module_count() {
-    let _ = env_logger::builder()
-        .is_test(true)
-        .format_timestamp(None) // Disable timestamps
-        .try_init();
-    let fixture_name = "file_dir_detection";
-    // Avoid tuple deconstruction
-    let graph_and_tree = build_tree_for_tests(fixture_name);
-    let graph = graph_and_tree.0;
-    let tree = graph_and_tree.1;
-
-    // Assert that the number of modules in the tree's map equals the number in the merged graph
-    assert_eq!(
-        tree.modules().len(),
-        graph.graph.modules.len(),
-        "ModuleTree should contain all modules from the merged graph"
-    );
-}
+// NOTE:
+// This test is replaced by unit testing in `parsed_graph.rs`,
+//  see parsed_graph::tests::test_build_mod_tree_inners
+// #[cfg(not(feature = "type_bearing_ids"))]
+// fn test_module_tree_module_count() {
+// }
 /// **Covers:** Correct population of the `path_index` field within the `ModuleTree`.
 /// It verifies that the canonical paths (e.g., `["crate"]`, `["crate", "top_pub_mod"]`,
 /// `["crate", "top_pub_mod", "nested_pub"]`, `["crate", "inline_pub_mod"]`) for
